backend/requirements.txt
langchain
langchain-openai
langchain-community
langchain-experimental
tavily-python
python-dotenv
requests
openai
playsound
SpeechRecognition
pyaudio
pillow
fastapi
uvicorn
